Introduce asserting invalid state (#45)
* Ensure use of static type AVClient
* Extract signing of votes to seperate function
* Recorded server responses to get new signatures etc.
* Use AV client internal state rather than call sequence recording

When detecting if an operation is invalid, relying on actual internal
state is better than checking whether a specific call sequence has
occurred. Otherwise we would  effectively be storing state that could
be easily derived from already existing state.

Mark two tests as skipped. They rely on an imcomplete implementation of
spoilBallotCryptograms().

* Remove successedMethods, used for recording call sequence

This has been replaced by relying on existing internal state.

* Temp cast to any to avoid type error

This will be replaced by a proper refactoring. In the mean time this
solves the build error of using a private state variable.

* Remove deprecated recorded server replies