Add inline comments on error handling
SJCL will throw errors if a caller attempts to instantiate a point using
an x-value, that does not have a corresponding y-value on the elliptic
curve.

We use the strategy of trying x+1, when this occurs. It also means that
we are "silently" catching an error and retrying another x-value.