Use new host names for authorizer and otp-provider during tests
To support testing AVX, VA and OTP provider running both natively and
using docker, we need the DNS names of all three to be consistent. This
commit fixes a problem that occurred when recording an election_config
to disk, which did not match the type of execution (docker vs native).

The means that all developers needs to ensure the following line is
present in their /etc/hosts file:
127.0.0.1 us-avx otp-provider authorizer