Add optional validation of payload content
Most items returned from the DBB should be validated for having the
expected content. However, the BoardEncryptionCommitmentItem will have
content that cannot be predicted. Thus content validation does not make
sense for this particular item.

To account for this, the validatePayload method now allows the "content"
part of the expectation-parameter to be undefined, meaning that a
content validation will not occur.